v2/requirements.txt
# requirements.txt
Flask
Flask-SQLAlchemy
Flask-JWT-Extended
psycopg2-binary
werkzeug
python-dateutil
Flask-Marshmallow
marshmallow-sqlalchemy